Refactor error msg on orderspage.html
